split fns
